% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppliMatPauli.R
\name{Amatrice_pauli}
\alias{Amatrice_pauli}
\title{onction qui applique une matrice de pauli pour la mesure}
\usage{
Amatrice_pauli(etat, matrix)
}
\arguments{
\item{etat}{vecteur d'etat initial}

\item{matrix}{pauli}
}
\value{
Un vecteur d'etat
}
\description{
onction qui applique une matrice de pauli pour la mesure
}
