% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EtatSuperposition.R
\name{etat_superposition}
\alias{etat_superposition}
\title{fonction qui cree l'état de superposition}
\usage{
etat_superposition(alpha, beta)
}
\arguments{
\item{alpha}{Nombre complexe ou réel}

\item{beta}{Nombre complexe ou réel}
}
\value{
Un vecteur d'état normalisé
}
\description{
fonction qui cree l'état de superposition
}
