% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbaMesure.R
\name{proba_mesure}
\alias{proba_mesure}
\title{Fonction qui calcule la probabilite de la mesure}
\usage{
proba_mesure(etat)
}
\arguments{
\item{etat}{Vecteur d'etat normalisé}
}
\value{
Un vecteur avec les Probabilites
}
\description{
Fonction qui calcule la probabilite de la mesure
}
