% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffMesure.R
\name{simul_mesure}
\alias{simul_mesure}
\title{Fonction qui simule la mesure}
\usage{
simul_mesure(etat)
}
\arguments{
\item{etat}{Vecteur etat initial}
}
\value{
Une liste avec resultat
}
\description{
Fonction qui simule la mesure
}
